ref: contained logic in its own component where it makes sense
