+++ src/pages/SignUp.tsx
@@ -1,8 +1,27 @@
 import AuthForm from "@/components/AuthForm";
+import { useState, useContext } from "react";
+import { FirebaseContext } from "@/App";
+
 const SignUp = () => {
+	const [error, setError] = useState<string>("");
+	const { db, auth } = useContext(FirebaseContext);
+	async function signUpUser(
+		e: React.InvalidEvent<HTMLFormElement>
+	): Promise<void> {
+		e.preventDefault();
+		const form = new FormData(e.currentTarget);
+		const formEntries = Object.fromEntries(form.entries());
+		console.log(formEntries);
+		console.log("signed up");
+	}
 	return (
 		<>
-			<AuthForm />
+			<AuthForm
+				setError={setError}
+				handleSubmit={signUpUser}
+				error={error}
+				action={window.location.pathname}
+			/>
 		</>
 	);
 };
